move the analysis functions from the AnalyzerNode to a seperate LogAnalyzer class
